bash update added to formsapp

+++ FormsApp/Controllers/HomeController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Diagnostics;
+using System.Reflection;
 
 namespace FormsApp.Controllers
 {
@@ -143,7 +144,34 @@ namespace FormsApp.Controllers
 			{
 				return NotFound();
 			}
+			return View("DeleteConfirm",entity);
+		}
+
+		[HttpPost]
+		public IActionResult Delete(int id,int ProductId)
+		{
+			if (id != ProductId)
+			{
+				return NotFound();
+			}
+			var entity = Repository.Products.FirstOrDefault(p => p.ProductId == ProductId);
+			if (entity == null)
+			{
+				return NotFound();
+			}
+
 			Repository.DeleteProduct(entity);
+			return RedirectToAction("Index");
+		}
+
+		[HttpPost]
+		public IActionResult EditProducts(List<Product> Products) 
+		{ 
+			foreach (var product in Products)
+			{
+				Repository.EditIsActive(product);
+			}
+
 			return RedirectToAction("Index");
 		}
